Moved removeTodo function to App.js
