+++ src/App.js
@@ -5,6 +5,7 @@ import TodoList from "./components/TodoList";
 import style from "./styles.module.css";
 import checklistIcon from "./svgs/checklist.svg";
 import initialFetchList, { createTodo } from "./components/Airtable";
+import PropTypes from "prop-types";
 
 function App() {
   //This is the list of todo items and a state checker to see if page is loading
@@ -16,14 +17,28 @@ function App() {
     initialFetchList(setTodoList, setIsLoading);
   }, []);
 
-  function addTodo(newTodo) {
-    createTodo(newTodo, todoList, setTodoList);
+  function addTodo(title) {
+    createTodo(title, todoList, setTodoList);
+  }
+
+  function removeTodo(id) {
+    fetch(`https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/default/${id}`, {
+      method: "DELETE",
+      headers: { Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_KEY}` },
+    }).then((response) => console.log("deleted todo list item"));
+
+    const filteredList = todoList.filter((item) => item.id !== id);
+    setTodoList(filteredList);
   }
 
   return (
     <Router>
       <Routes>
-        <Route path="/" exact element={<Home isLoading={isLoading} todoList={todoList} setTodoList={setTodoList} onAddTodo={addTodo} />} />
+        <Route
+          path="/"
+          exact
+          element={<Home isLoading={isLoading} todoList={todoList} setTodoList={setTodoList} onAddTodo={addTodo} onRemoveTodo={removeTodo} />}
+        />
         <Route path="/new" exact element={<New />} />
       </Routes>
     </Router>
@@ -31,7 +46,7 @@ function App() {
 }
 
 const Home = (props) => {
-  const { isLoading, todoList, setTodoList, onAddTodo } = props;
+  const { isLoading, todoList, setTodoList, onAddTodo, onRemoveTodo } = props;
   return (
     <div className={style.container}>
       <div className={style.header}>
@@ -43,7 +58,7 @@ const Home = (props) => {
         <h1>Todo List</h1>
       </div>
       <AddTodoForm todoList={todoList} setTodoList={setTodoList} onAddTodo={onAddTodo} />
-      {isLoading ? <p>Loading...</p> : <TodoList todoList={todoList} setTodoList={setTodoList} />}
+      {isLoading ? <p>Loading...</p> : <TodoList todoList={todoList} setTodoList={setTodoList} onRemoveTodo={onRemoveTodo} />}
     </div>
   );
 };
@@ -76,6 +91,20 @@ const New = () => {
       </ul>
     </>
   );
+};
+
+App.propTypes = {
+  todoList: PropTypes.array,
+  setTodoList: PropTypes.func,
+  isLoading: PropTypes.bool,
+  setIsLoading: PropTypes.func,
+};
+
+Home.propTypes = {
+  isLoading: PropTypes.bool,
+  todoList: PropTypes.array,
+  setTodoList: PropTypes.func,
+  onAddTodo: PropTypes.func,
 };
 
 export default App;
