+++ src/components/TodoListItem.js
@@ -4,22 +4,12 @@ import style from "./TodoListItem.module.css";
 import PropTypes from "prop-types";
 
 function TodoListItem(props) {
-  const { item, todoList, setTodoList } = props;
-
-  function removeTodo(id) {
-    fetch(`https://api.airtable.com/v0/${process.env.REACT_APP_AIRTABLE_BASE_ID}/default/${id}`, {
-      method: "DELETE",
-      headers: { Authorization: `Bearer ${process.env.REACT_APP_AIRTABLE_API_KEY}` },
-    }).then((response) => console.log("deleted todo list item"));
-
-    const filteredList = todoList.filter((item) => item.id !== id);
-    setTodoList(filteredList);
-  }
+  const { item, onRemoveTodo } = props;
 
   return (
     <li className={style.ListItem}>
       {item.fields.Title}
-      <button className={stylez.removeButton} onClick={() => removeTodo(item.id)}>
+      <button className={stylez.removeButton} onClick={() => onRemoveTodo(item.id)}>
         X
       </button>
     </li>
@@ -28,8 +18,7 @@ function TodoListItem(props) {
 
 TodoListItem.propTypes = {
   item: PropTypes.object,
-  todoList: PropTypes.array,
-  setTodoList: PropTypes.func,
+  onRemoveTodo: PropTypes.func,
 };
 
 export default TodoListItem;
